archive/legacy-desktop-pyqt/archive/requirements_rembg_gpu.txt
rembg==2.0.59
Pillow==10.4.0
onnxruntime-gpu==1.18.1
